authorizer-lambda/requirements.txt
PyJWT
cryptography
python-jose
